refacto: rename GamesCarousel and insert Carousel config
