+++ src/components/games/GamesCarousel.styles.ts
