+++ src/styles/index.ts
@@ -2,6 +2,5 @@ import * as Colors from './colors'
 import * as Spacing from './spacing'
 import * as Percentage from './percentage'
 import * as Font from './font'
-import * as Carousel from './carousel'
 
-export { Colors, Spacing, Percentage, Font, Carousel }
+export { Colors, Spacing, Percentage, Font }
